Work towards 3.1.2
reversioned to 3.1.2 - do not plan on having major new features in next
release. The changelog has been altered to reflect this.
Logger appends message to [PluginName], no need for us to do that.
Added a start at /book usage - will have to redo so that people without
permission can't see usages they're not supposed to with .contains
autocompletion